+++ src/Data/TypedEncoding/Conv/ByteString/Lazy/Char8.hs
@@ -34,3 +34,23 @@ unpack :: (
         , AllEncodeInto "r-CHAR8" encs
        ) => Enc xs c BL8.ByteString -> Enc xs c String
 unpack = unsafeChangePayload BL8.unpack          
+
+-- | 
+-- Lazy version of 'Data.TypedEncoding.Conv.ByteString.Char8.pack'''.
+pack'' :: (
+    Knds.UnSnoc xs ~ '(,) ys y
+    , EncodingAnn y
+    , encs ~ RemoveRs ys
+    , AllEncodeInto "r-CHAR8" encs
+    ) => Enc xs c String -> Enc xs c BL8.ByteString
+pack'' = unsafeChangePayload BL8.pack
+
+-- | 
+-- Lazy version of 'Data.TypedEncoding.Conv.ByteString.Char8.unpack'''.
+unpack'' :: (
+          Knds.UnSnoc xs ~ '(,) ys y
+         , EncodingAnn y
+         , encs ~ RemoveRs ys
+         , AllEncodeInto "r-CHAR8" encs
+          ) => Enc xs c BL8.ByteString -> Enc xs c String
+unpack'' = unsafeChangePayload BL8.unpack